+++ lib/august_db.ex
@@ -6,6 +6,4 @@ defmodule AugustDb do
   Contexts are also responsible for managing your data, regardless
   if it comes from the database, an external API or others.
   """
-
-  use Memtable
 end
